Rename initialize.py to app_initializer.py
